Moved JsonCodec interfaces to io.vertx.core.spi.json
