+++ src/test/java/io/vertx/test/codegen/testjsoncodecs/enclosedcodec/MyPojo.java
@@ -1,6 +1,6 @@
 package io.vertx.test.codegen.testjsoncodecs.enclosedcodec;
 
-import io.vertx.core.json.JsonCodec;
+import io.vertx.core.spi.json.JsonCodec;
 import io.vertx.core.json.JsonObject;
 
 public class MyPojo {
