+++ src/test/java/io/vertx/test/codegen/testjsoncodecs/illegaljsontypeincodec/ZonedDateTimeCodec.java
@@ -1,6 +1,6 @@
 package io.vertx.test.codegen.testjsoncodecs.illegaljsontypeincodec;
 
-import io.vertx.core.json.JsonCodec;
+import io.vertx.core.spi.json.JsonCodec;
 
 import java.time.Instant;
 import java.time.ZonedDateTime;
